refactor: file folder location for all data routes
