+++ src/server.ts
@@ -2,7 +2,7 @@ import express from "express";
 import swaggerUI from "swagger-ui-express";
 import swaggerJSDocs from "swagger-jsdoc";
 import { db, swaggerConfig } from "./config";
-import InformationsRoutes from "./api/information/routes/information.routes";
+import InformationsRoutes from "./api/information/information.routes";
 import { createBanner } from "./api/information/entities/banner.entity";
 import { createService } from "./api/information/entities/service.entity";
 import "dotenv/config";
